refactor(github-issues): extract shared hash helpers
