+++ crates/tau-github-issues/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2021"
 [dependencies]
 reqwest = { workspace = true }
 chrono = { workspace = true }
+sha2 = { workspace = true }
